Making services not start by default
The idea behind this change is that you will want Wick to install and
configure all sorts of software, but not necessarily start the services.
Perhaps you are building a machine that will be immediately imaged and
cloned.  Or you are dealing with a service that requires frequent
restarts when you are configuring it fully.

By putting this decision into the hands of the sysop who is writing the
other formulas and roles, we let them decide how they want the machine
brought up.

The `wickService force-state` will always leave the machine in a
consistent state, regardless if the service was installed and running
previously.  Idempotent scripts are good.