Only update /etc/resolv.conf with --start + fix the update
Before this always updated /etc/resolv.conf, which should not happen if
the service does not start because then all DNS lookups would hang.
Also, the sed was wrong and would potentially place multiple nameserver
entries in the file.

This new revision doesn't modify the file in-place but it does at least
fix the problems mentioned above.

Thanks for @baelish for working with me on this one.  This closes #33.