Fixing undefined variable
This happened when I cleaned up the script to match my coding style
guide.  I renamed a variable elsewhere but then did not notice I used it
globally elsewhere.  This has now been fixed in two ways - first it is
not global and secondly the value is passed properly and the right
variable name is now used.