Machine type order (#58)
* changing method of retreiving version for ubuntu

The current method was giving me "LTS" which was not desired.  Also
it was using `lsb-release` instead of `lsb_release`.

* May as well match as much as I can

In case "Release" shows up in one of the other fields values

* Another attempt

With a slimmer version of ubuntu lsb_release didn't exist but the
file /etc/lsb-release should always exist

* changed to try different approaches

* checking docker first

because it is very quick to check.  Amazon takes a long time before
it fails.