Systemd services (#63)
* adding a few of the initial changes for systemd

* handling some of the override

I think I could more thoroughly handle this, and I might have to
more thoroughly handle this

* more specialized now

* adding progress

There is now a split between upstart and systemd for adding
dependencies.

* reimplementing make-override

reluctantly.  It does have a use

* fixing variable name

* Adding ability to check if a service is enabled.

This is a work in progress but I wanted it backed up somewhere.  I
plan to use this to determine if I need to depend on additional
services in more specific scenarios.

* Removing upstart functionality

This looks like a bit of a larger problem.  What we support right now
is sysv script that are running in compatibility mode and now systemd.

upstart services are not supported.  We would have to add them
differently and to a different location and change enabling
functionality.

* adding some doc blocks

* adding to wickService

* ability to add if it exists

So that I can add a very common one liner in many of our formulas

* ability to add if it exists

So that I can add a very common one liner in many of our formulas

(cherry picked from commit 547256a184f088f08cf797ae3b0a056cce975ac6)

* adding some doc blocks

* making executable

to match the rest of the files

* fixing filename

* adding doc block

* renaming files

Next comes renaming inside files.  Making this change because these
are actually sysv services.

* renaming inside files.

Also corrected a mistake in a sed command.

* missed this one

* Updating docs

* renaming

Because it calls is-enabled

* Fixing various issues

All of them small.

* fixing style and doc blocks

* updating generated docs

Forgot to do this before last commit

* so it doesn't output

* alphabetizing

* no longer needed