Fix a couple bugs
If a username was provided or if a destination was not provided it
would blow up because "args" was defined but not assigned as an
array.  destination tried to default to "-" but it failed because
it used the wrong default syntax (where it would default only if it
was not defined, which it was).

Further, args was never used anyways.

Also some of the docs were wrong