Updating code to work with newer versions of Bash
As it mentions in doc/contexts-that-disable-exit-on-error, this is a
problem that comes with newer versions of Bash.  It seems to appear
somewhere on or just before Bash 4.3.

It's a LOT of updates because I also ran ShellCheck against all of the
shell scripts, updated examples, and fixed things that were not up to
the Tests Always Included style guide.